Modify item adding code to store cover_photo

+++ src/routes/newItem.js
@@ -54,7 +54,7 @@ const handler = (pool) => {
         async (req, res) => {
             const errors = validationResult(req);
             if (!errors.isEmpty()) {
-                if (req.files) {
+                if (req.files && Object.keys(req.files).length !== 0) {
                     Object.keys(req.files).forEach((key) => {
                         req.files[key].forEach(async (image) => {
                             await deleteFile(image.path);
@@ -69,13 +69,24 @@ const handler = (pool) => {
             const { name, description, price } = req.body;
             const itemId = randomUUID();
             let coverPhotoPath;
-            let imagePaths; 
-            if (req.files) {
-                coverPhotoPath = req.files["cover-photo"][0].filename;
-                imagePaths = req.files["images"].map((image) => image.filename);
+            let imagePaths;
+            if (req.files && Object.keys(req.files).length !== 0) {
+                if (req.files["cover-photo"])
+                    coverPhotoPath = req.files["cover-photo"][0].filename;
+                if (req.files["images"])
+                    imagePaths = req.files["images"].map(
+                        (image) => image.filename
+                    );
             }
             try {
-                await addItem(pool, itemId, name, description, price);
+                await addItem(
+                    pool,
+                    itemId,
+                    name,
+                    description,
+                    price,
+                    coverPhotoPath ? coverPhotoPath : null
+                );
                 if (coverPhotoPath) {
                     await addImage(pool, itemId, coverPhotoPath);
                 }
